+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 * Husk of the Pit will no longer be equipped by the Item Leveling loadout.
 * Fixed equipping loadouts onto the current character from Loadout Builder.
 * The default shader no longer counts as a duplicate item.
+* DIM no longer tries to equip exotic faction class items where your character isn't aligned with the right faction.
 
 # 3.10.6